refactor!: drop the served-actions seam from mount runtimes
The socket that served a mount's action registry is gone (ADR-0112), so
nothing reads DaemonRuntime.actions: the watcher only disposes the
runtime and reports collaboration status. Remove the field, delete
MountComposition, and make compose side-effect only; it exists to attach
materializers, which enroll their own drain through scope.registerDrain.

The e2e fixture loses its counter actions (their one purpose was giving
'epicenter run demo.counter_*' something to call) and the stub
collaboration fields the Collaboration type never had; the fixture dir
is renamed inline-actions -> demo-mount to match.
