+++ packages/workspace/src/daemon/types.ts
@@ -1,17 +1,16 @@
 /**
  * Daemon-side runtime types.
  *
- * `DaemonRuntime` is the contract every opened mount returns: async dispose
- * plus the local action registry available inside the watcher process.
- * Collaborative mounts may also expose a hosted `Collaboration` for identity,
- * sync, and peer presence.
+ * `DaemonRuntime` is the contract every opened mount returns: async dispose,
+ * plus an optional hosted `Collaboration` for identity, sync, and peer
+ * presence. The watcher is not a callable action server (ADR-0112), so the
+ * runtime carries no action registry.
  *
  * `StartedMount` is the lifecycle-owning mount shape opened from a configured
  * mount factory.
  */
 
 import type { Collaboration } from '../document/open-collaboration.js';
-import type { ActionRegistry } from '../shared/actions.js';
 import type { MaybePromise } from '../shared/types.js';
 
 /**
@@ -20,9 +19,6 @@ import type { MaybePromise } from '../shared/types.js';
 export type DaemonRuntime = {
 	/** Called by the daemon at exit. */
 	[Symbol.asyncDispose](): MaybePromise<void>;
-
-	/** The action registry available inside this runtime. */
-	readonly actions: ActionRegistry;
 
 	/**
 	 * Optional hosted collaboration. Identity, sync status, and live-node
